+++ lib/SlimViews/TwigExtension.php
@@ -1,0 +1,92 @@
+<?php
+/**
+ * Slim - a micro PHP 5 framework
+ *
+ * @author      Josh Lockhart
+ * @author      Andrew Smith
+ * @link        http://www.slimframework.com
+ * @copyright   2013 Josh Lockhart
+ * @version     0.1.3
+ * @package     SlimViews
+ *
+ * MIT LICENSE
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+namespace Slim\Views;
+
+use Slim\Slim;
+
+class TwigExtension extends \Twig\Extension\AbstractExtension
+{
+    public function getName()
+    {
+        return 'slim';
+    }
+
+    public function getFunctions()
+    {
+        return [
+            new \Twig\TwigFunction('urlFor', [$this, 'urlFor']),
+            new \Twig\TwigFunction('baseUrl', [$this, 'base']),
+            new \Twig\TwigFunction('siteUrl', [$this, 'site']),
+            new \Twig\TwigFunction('currentUrl', [$this, 'currentUrl']),
+        ];
+    }
+
+    public function urlFor($name, $params = [], $appName = 'default')
+    {
+        return Slim::getInstance($appName)->urlFor($name, $params);
+    }
+
+    public function site($url, $withUri = true, $appName = 'default')
+    {
+        return $this->base($withUri, $appName) . '/' . ltrim($url, '/');
+    }
+
+    public function base($withUri = true, $appName = 'default')
+    {
+        $req = Slim::getInstance($appName)->request();
+        $uri = $req->getUrl();
+
+        if ($withUri) {
+            $uri .= $req->getRootUri();
+        }
+        return $uri;
+    }
+
+    public function currentUrl($withQueryString = true, $appName = 'default')
+    {
+        $app = Slim::getInstance($appName);
+        $req = $app->request();
+        $uri = $req->getUrl() . $req->getPath();
+
+        if ($withQueryString) {
+            $env = $app->environment();
+
+            if ($env['QUERY_STRING']) {
+                $uri .= '?' . $env['QUERY_STRING'];
+            }
+        }
+
+        return $uri;
+    }
+}